+++ packages/orq-rc/RELEASES.md
@@ -6719,3 +6719,13 @@ Based on:
 - [typescript v4.3.0-rc.58] packages/orq-rc
 ### Releases
 - [NPM v4.3.0-rc.58] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.58 - packages/orq-rc
+
+## 2026-02-12 11:59:17
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.713.0 (2.821.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.3.0-rc.59] packages/orq-rc
+### Releases
+- [NPM v4.3.0-rc.59] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.59 - packages/orq-rc